Moves welcome Page to inventory
Removes menu if not logged in (in inventory and create)

Adds no chat selected info instead of welcome in chats
