+++ webofneeds/won-owner-webapp/src/main/webapp/style/_connections.scss
@@ -10,35 +10,23 @@ aside.overview__left {
   }
 }
 
-main.overview__right {
-  .overview__right__welcome {
-    padding: 0.5rem;
+main.overview__rightwelcome {
+  .overview__rightwelcome__noselection {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background: $won-lighter-gray;
 
-    > won-usecase-picker {
-      padding: 0;
+    &__icon {
+      @include fixed-square($hugeiconSize);
+      --local-primary: #{$won-line-gray};
     }
 
     &__text {
-      padding-bottom: 1rem;
-      h1 {
-        padding-bottom: 1.5rem;
-        text-align: center;
-      }
-      p {
-        padding-bottom: 1.5rem;
-        text-align: center;
-      }
-      h3 {
-        text-align: center;
-      }
-    }
-
-    &__howto {
-      padding-top: 3rem;
-
-      &__button {
-        width: 100%;
-      }
+      margin-top: 1rem;
+      color: $won-line-gray;
+      font-size: $mediumFontSize;
     }
   }
 }
